Reference

Reference

Types

EzXML.Document — Type.

An XML/HTML document type.

source

EzXML.Node — Type.

A proxy type to libxml2's node struct.

Properties (Julia ≥ 0.7)

Name	Type	Description

type	EzXML.NodeType	the type of a node

name	String?	the name of a node

path	String	the absolute path to a node

content	String	the content of a node

namespace	String?	the namespace associated with a node

source

EzXML.StreamReader — Type.

A streaming XML reader type.

source

EzXML.XMLError — Type.

An error detected by libxml2.

source

I/O

EzXML.parsexml — Function.

parsexml(xmlstring)

Parse xmlstring and create an XML document.

source

EzXML.parsehtml — Function.

parsehtml(htmlstring)

Parse htmlstring and create an HTML document.

source

EzXML.readxml — Function.

readxml(filename)

Read filename and create an XML document.

source

readxml(input::IO)

Read input and create an XML document.

source

EzXML.readhtml — Function.

readhtml(filename)

Read filename and create an HTML document.

source

readhtml(input::IO)

Read input and create an HTML document.

source

EzXML.prettyprint — Function.

prettyprint([io], node::Node)

Print node with formatting.

source

prettyprint([io], doc::Document)

Print doc with formatting.

source

Constructors

EzXML.XMLDocument — Function.

XMLDocument(version="1.0")

Create an XML document.

source

EzXML.HTMLDocument — Function.

HTMLDocument(uri=nothing, externalID=nothing)

Create an HTML document.

source

EzXML.XMLDocumentNode — Function.

XMLDocumentNode(version)

Create an XML document node with version.

source

EzXML.HTMLDocumentNode — Function.

HTMLDocumentNode(uri, externalID)

Create an HTML document node.

uri and externalID are either a string or nothing.

source

EzXML.ElementNode — Function.

ElementNode(name)

Create an element node with name.

source

EzXML.TextNode — Function.

TextNode(content)

Create a text node with content.

source

EzXML.CommentNode — Function.

CommentNode(content)

Create a comment node with content.

source

EzXML.CDataNode — Function.

CDataNode(content)

Create a CDATA node with content.

source

EzXML.AttributeNode — Function.

AttributeNode(name, value)

Create an attribute node with name and value.

source

EzXML.DTDNode — Function.

DTDNode(name, [systemID, [externalID]])

Create a DTD node with name, systemID, and externalID.

source

Node types

Node type	Integer

EzXML.ELEMENT_NODE	1

EzXML.ATTRIBUTE_NODE	2

EzXML.TEXT_NODE	3

EzXML.CDATA_SECTION_NODE	4

EzXML.ENTITY_REF_NODE	5

EzXML.ENTITY_NODE	6

EzXML.PI_NODE	7

EzXML.COMMENT_NODE	8

EzXML.DOCUMENT_NODE	9

EzXML.DOCUMENT_TYPE_NODE	10

EzXML.DOCUMENT_FRAG_NODE	11

EzXML.NOTATION_NODE	12

EzXML.HTML_DOCUMENT_NODE	13

EzXML.DTD_NODE	14

EzXML.ELEMENT_DECL	15

EzXML.ATTRIBUTE_DECL	16

EzXML.ENTITY_DECL	17

EzXML.NAMESPACE_DECL	18

EzXML.XINCLUDE_START	19

EzXML.XINCLUDE_END	20

EzXML.DOCB_DOCUMENT_NODE	21

Node accessors

EzXML.nodetype — Method.

nodetype(node::Node)

Return the type of node as an integer.

source

EzXML.nodepath — Method.

nodepath(node::Node)

Return the path of node.

source

EzXML.nodename — Method.

nodename(node::Node)

Return the node name of node.

source

EzXML.nodecontent — Method.

nodecontent(node::Node)

Return the node content of node.

source

EzXML.namespace — Method.

namespace(node::Node)

Return the namespace associated with node.

source

EzXML.namespaces — Method.

namespaces(node::Node)

Create a vector of namespaces applying to node.

source

EzXML.hasnodename — Method.

hasnodename(node::Node)

Return if node has a node name.

source

EzXML.hasnamespace — Method.

hasnamespace(node::Node)

Return if node is associated with a namespace.

source

EzXML.iselement — Method.

iselement(node::Node)

Return if node is an element node.

source

EzXML.isattribute — Method.

isattribute(node::Node)

Return if node is an attribute node.

source

EzXML.istext — Method.

istext(node::Node)

Return if node is a text node.

source

EzXML.iscdata — Method.

iscdata(node::Node)

Return if node is a CDATA node.

source

EzXML.iscomment — Method.

iscomment(node::Node)

Return if node is a comment node.

source

EzXML.isdtd — Method.

isdtd(node::Node)

Return if node is a DTD node.

source

EzXML.countnodes — Method.

countnodes(parent::Node)

Count the number of child nodes of parent.

source

EzXML.countelements — Method.

countelements(parent::Node)

Count the number of child elements of parent.

source

EzXML.countattributes — Method.

countattributes(elem::Node)

Count the number of attributes of elem.

source

EzXML.systemID — Method.

systemID(node::Node)

Return the system ID of node.

source

EzXML.externalID — Method.

externalID(node::Node)

Return the external ID of node.

source

Node modifiers

EzXML.setnodename! — Method.

setnodename!(node::Node, name::AbstractString)

Set the name of node.

source

EzXML.setnodecontent! — Method.

setnodecontent!(node::Node, content::AbstractString)

Replace the content of node.

source

Document properties

EzXML.version — Function.

version(doc::Document)

Return the version string of doc.

source

EzXML.encoding — Function.

encoding(doc::Document)

Return the encoding string of doc.

source

EzXML.hasversion — Function.

hasversion(doc::Document)

Return if doc has a version string.

source

EzXML.hasencoding — Function.

hasencoding(doc::Document)

Return if doc has an encoding string.

source

DOM tree accessors

EzXML.document — Function.

document(node::Node)

Return the document of node.

source

EzXML.root — Function.

root(doc::Document)

Return the root element of doc.

source

EzXML.dtd — Function.

dtd(doc::Document)

Return the DTD node of doc.

source

EzXML.parentnode — Function.

parentnode(node::Node)

Return the parent of node.

source

EzXML.parentelement — Function.

parentelement(node::Node)

Return the parent element of node.

source

EzXML.firstnode — Function.

firstnode(node::Node)

Return the first child node of node.

source

EzXML.lastnode — Function.

lastnode(node::Node)

Return the last child node of node.

source

EzXML.firstelement — Function.

firstelement(node::Node)

Return the first child element of node.

source

EzXML.lastelement — Function.

lastelement(node::Node)

Return the last child element of node.

source

EzXML.nextnode — Function.

nextnode(node::Node)

Return the next node of node.

source

EzXML.prevnode — Function.

prevnode(node::Node)

Return the previous node of node.

source

EzXML.nextelement — Function.

nextelement(node::Node)

Return the next element of node.

source

EzXML.prevelement — Function.

prevelement(node::Node)

Return the previous element of node.

source

EzXML.eachnode — Function.

eachnode(node::Node, [backward=false])

Create an iterator of child nodes.

source

EzXML.nodes — Function.

nodes(node::Node, [backward=false])

Create a vector of child nodes.

source

EzXML.eachelement — Function.

eachelement(node::Node, [backward=false])

Create an iterator of child elements.

source

EzXML.elements — Function.

elements(node::Node, [backward=false])

Create a vector of child elements.

source

EzXML.eachattribute — Function.

eachattribute(node::Node)

Create an iterator of attributes.

source

eachattribute(reader::StreamReader)

Return an AttributeReader object for the current node of reader

source

EzXML.attributes — Function.

attributes(node::Node)

Create a vector of attributes.

source

EzXML.hasroot — Function.

hasroot(doc::Document)

Return if doc has a root element.

source

EzXML.hasdtd — Function.

hasdtd(doc::Document)

Return if doc has a DTD node.

source

EzXML.hasnode — Function.

hasnode(node::Node)

Return if node has a child node.

source

EzXML.hasnextnode — Function.

hasnextnode(node::Node)

Return if node has a next node.

source

EzXML.hasprevnode — Function.

hasprevnode(node::Node)

Return if node has a previous node.

source

EzXML.haselement — Function.

haselement(node::Node)

Return if node has a child element.

source

EzXML.hasnextelement — Function.

hasnextelement(node::Node)

Return if node has a next node.

source

EzXML.hasprevelement — Function.

hasprevelement(node::Node)

Return if node has a previous node.

source

EzXML.hasdocument — Function.

hasdocument(node::Node)

Return if node belongs to a document.

source

EzXML.hasparentnode — Function.

hasparentnode(node::Node)

Return if node has a parent node.

source

EzXML.hasparentelement — Function.

hasparentelement(node::Node)

Return if node has a parent node.

source

DOM tree modifiers

EzXML.setroot! — Function.

setroot!(doc::Document, node::Node)

Set the root element of doc to node and return the root element.

source

EzXML.setdtd! — Function.

setdtd!(doc::Document, node::Node)

Set the DTD node of doc to node and return the DTD node.

source

EzXML.link! — Function.

link!(parent::Node, child::Node)

Link child at the end of children of parent.

source

EzXML.linknext! — Function.

linknext!(target::Node, node::Node)

Link node as the next sibling of target.

source

EzXML.linkprev! — Function.

linkprev!(target::Node, node::Node)

Link node as the prev sibling of target.

source

EzXML.unlink! — Function.

unlink!(node::Node)

Unlink node from its context.

source

EzXML.addelement! — Function.

addelement!(parent::Node, name::AbstractString)

Add a new child element of name with no content to parent and return the new child element.

source

addelement!(parent::Node, name::AbstractString, content::AbstractString)

Add a new child element of name with content to parent and return the new child element.

source

XPath query

Base.findall — Method.

findall(xpath::AbstractString, doc::Document)

Find nodes matching xpath XPath query from doc.

source

Base.findfirst — Method.

findfirst(xpath::AbstractString, doc::Document)

Find the first node matching xpath XPath query from doc.

source

Base.findlast — Method.

findlast(doc::Document, xpath::AbstractString)

Find the last node matching xpath XPath query from doc.

source

Base.findall — Method.

findall(xpath::AbstractString, node::Node, [ns=namespaces(node)])

Find nodes matching xpath XPath query starting from node.

The ns argument is an iterator of namespace prefix and URI pairs.

source

Base.findfirst — Method.

findfirst(xpath::AbstractString, node::Node, [ns=namespaces(node)])

Find the first node matching xpath XPath query starting from node.

source

Base.findlast — Method.

findlast(node::Node, xpath::AbstractString, [ns=namespaces(node)])

Find the last node matching xpath XPath query starting from node.

source

Validation

EzXML.validate — Function.

validate(doc::Document, [dtd::Node])

Validate doc against dtd and return the validation log.

The validation log is empty if and only if doc is valid. The DTD node in doc will be used if dtd is not passed.

source

EzXML.readdtd — Function.

readdtd(filename::AbstractString)

Read filename and create a DTD node.

source

Reader node types

Node type	Integer

EzXML.READER_NONE	0

EzXML.READER_ELEMENT	1

EzXML.READER_ATTRIBUTE	2

EzXML.READER_TEXT	3

EzXML.READER_CDATA	4

EzXML.READER_ENTITY_REFERENCE	5

EzXML.READER_ENTITY	6

EzXML.READER_PROCESSING_INSTRUCTION	7

EzXML.READER_COMMENT	8

EzXML.READER_DOCUMENT	9

EzXML.READER_DOCUMENT_TYPE	10

EzXML.READER_DOCUMENT_FRAGMENT	11

EzXML.READER_NOTATION	12

EzXML.READER_WHITESPACE	13

EzXML.READER_SIGNIFICANT_WHITESPACE	14

EzXML.READER_END_ELEMENT	15

EzXML.READER_END_ENTITY	16

EzXML.READER_XML_DECLARATION	17

Streaming reader

EzXML.expandtree — Method.

expandtree(reader::StreamReader)

Expand the current node of reader into a full subtree that will be available until the next read of node.

Note that the expanded subtree is a read-only and temporary object. You cannot modify it or keep references to any nodes of it after reading the next node.

Currently, namespace functions and XPath query will not work on the expanded subtree.

source

EzXML.nodetype — Method.

nodetype(reader::StreamReader)

Return the type of the current node of reader.

source

EzXML.nodename — Method.

nodename(reader::StreamReader)

Return the name of the current node of reader.

source

EzXML.nodecontent — Method.

nodecontent(reader::StreamReader)

Return the content of the current node of reader.

source

EzXML.nodedepth — Method.

nodedepth(reader::StreamReader)

Return the depth of the current node of reader.

source

EzXML.namespace — Method.

namespace(reader::StreamReader)

Return the namespace of the current node of reader.

source

EzXML.hasnodecontent — Method.

hasnodecontent(reader::StreamReader)

Return if the current node of reader has content.

source

EzXML.hasnodename — Method.

hasnodename(reader::StreamReader)

Return if the current node of reader has a node name.

source