Reference

Types

EzXML.Document — Type

An XML/HTML document type.

EzXML.Node — Type

A proxy type to libxml2's node struct.

Properties

Name	Type	Description

type	EzXML.NodeType	the type of a node

name	String?	the name of a node

path	String	the absolute path to a node

content	String	the content of a node

namespace	String?	the namespace associated with a node

EzXML.StreamReader — Type

A streaming XML reader type.

EzXML.XMLError — Type

An error detected by libxml2.

I/O

EzXML.parsexml — Function

parsexml(xmlstring; options...)

Parse xmlstring and create an XML document.

Parsing Options

noerror = false: suppress (recoverable) error reports

nowarning = false: suppress warning reports

pedantic = false: pedantic error reporting

noblanks = false: remove blank nodes

EzXML.parsehtml — Function

parsehtml(htmlstring; options...)

Parse htmlstring and create an HTML document.

See parsexml for parsing options.

EzXML.readxml — Function

readxml(filename; options...)

Read filename and create an XML document.

See parsexml for parsing options.

readxml(input::IO; options...)

Read input and create an XML document.

See parsexml for parsing options.

EzXML.readhtml — Function

readhtml(filename; options...)

Read filename and create an HTML document.

See parsexml for parsing options.

readhtml(input::IO; options...)

Read input and create an HTML document.

See parsexml for parsing options.

EzXML.prettyprint — Function

prettyprint([io], node::Node)

Print node with formatting.

prettyprint([io], doc::Document)

Print doc with formatting.

Constructors

EzXML.XMLDocument — Function

XMLDocument(version="1.0")

Create an XML document.

EzXML.HTMLDocument — Function

HTMLDocument(uri=nothing, externalID=nothing)

Create an HTML document.

EzXML.XMLDocumentNode — Function

XMLDocumentNode(version)

Create an XML document node with version.

EzXML.HTMLDocumentNode — Function

HTMLDocumentNode(uri, externalID)

Create an HTML document node.

uri and externalID are either a string or nothing.

EzXML.ElementNode — Function

ElementNode(name)

Create an element node with name.

EzXML.TextNode — Function

TextNode(content)

Create a text node with content.

EzXML.CommentNode — Function

CommentNode(content)

Create a comment node with content.

EzXML.CDataNode — Function

CDataNode(content)

Create a CDATA node with content.

EzXML.AttributeNode — Function

AttributeNode(name, value)

Create an attribute node with name and value.

EzXML.DTDNode — Function

DTDNode(name, [systemID, [externalID]])

Create a DTD node with name, systemID, and externalID.

Node types

Node type	Integer

EzXML.ELEMENT_NODE	1

EzXML.ATTRIBUTE_NODE	2

EzXML.TEXT_NODE	3

EzXML.CDATA_SECTION_NODE	4

EzXML.ENTITY_REF_NODE	5

EzXML.ENTITY_NODE	6

EzXML.PI_NODE	7

EzXML.COMMENT_NODE	8

EzXML.DOCUMENT_NODE	9

EzXML.DOCUMENT_TYPE_NODE	10

EzXML.DOCUMENT_FRAG_NODE	11

EzXML.NOTATION_NODE	12

EzXML.HTML_DOCUMENT_NODE	13

EzXML.DTD_NODE	14

EzXML.ELEMENT_DECL	15

EzXML.ATTRIBUTE_DECL	16

EzXML.ENTITY_DECL	17

EzXML.NAMESPACE_DECL	18

EzXML.XINCLUDE_START	19

EzXML.XINCLUDE_END	20

EzXML.DOCB_DOCUMENT_NODE	21

Node accessors

EzXML.nodetype — Method

nodetype(node::Node)

Return the type of node as an integer.

EzXML.nodepath — Method

nodepath(node::Node)

Return the path of node.

EzXML.nodename — Method

nodename(node::Node)

Return the node name of node.

EzXML.nodecontent — Method

nodecontent(node::Node)

Return the node content of node.

EzXML.namespace — Method

namespace(node::Node)

Return the namespace associated with node.

EzXML.namespaces — Method

namespaces(node::Node)

Create a vector of namespaces applying to node.

EzXML.hasnodename — Method

hasnodename(node::Node)

Return if node has a node name.

EzXML.hasnamespace — Method

hasnamespace(node::Node)

Return if node is associated with a namespace.

EzXML.iselement — Method

iselement(node::Node)

Return if node is an element node.

EzXML.isattribute — Method

isattribute(node::Node)

Return if node is an attribute node.

EzXML.istext — Method

istext(node::Node)

Return if node is a text node.

EzXML.iscdata — Method

iscdata(node::Node)

Return if node is a CDATA node.

EzXML.iscomment — Method

iscomment(node::Node)

Return if node is a comment node.

EzXML.isdtd — Method

isdtd(node::Node)

Return if node is a DTD node.

EzXML.countnodes — Method

countnodes(parent::Node)

Count the number of child nodes of parent.

EzXML.countelements — Method

countelements(parent::Node)

Count the number of child elements of parent.

EzXML.countattributes — Method

countattributes(elem::Node)

Count the number of attributes of elem.

EzXML.systemID — Method

systemID(node::Node)

Return the system ID of node.

EzXML.externalID — Method

externalID(node::Node)

Return the external ID of node.

Node modifiers

EzXML.setnodename! — Method

setnodename!(node::Node, name::AbstractString)

Set the name of node.

EzXML.setnodecontent! — Method

setnodecontent!(node::Node, content::AbstractString)

Replace the content of node.

Document properties

EzXML.version — Function

version(doc::Document)

Return the version string of doc.

EzXML.encoding — Function

encoding(doc::Document)

Return the encoding string of doc.

EzXML.hasversion — Function

hasversion(doc::Document)

Return if doc has a version string.

EzXML.hasencoding — Function

hasencoding(doc::Document)

Return if doc has an encoding string.

DOM tree accessors

EzXML.document — Function

document(node::Node)

Return the document of node.

EzXML.root — Function

root(doc::Document)

Return the root element of doc.

EzXML.dtd — Function

dtd(doc::Document)

Return the DTD node of doc.

EzXML.parentnode — Function

parentnode(node::Node)

Return the parent of node.

EzXML.parentelement — Function

parentelement(node::Node)

Return the parent element of node.

EzXML.firstnode — Function

firstnode(node::Node)

Return the first child node of node.

EzXML.lastnode — Function

lastnode(node::Node)

Return the last child node of node.

EzXML.firstelement — Function

firstelement(node::Node)

Return the first child element of node.

EzXML.lastelement — Function

lastelement(node::Node)

Return the last child element of node.

EzXML.nextnode — Function

nextnode(node::Node)

Return the next node of node.

EzXML.prevnode — Function

prevnode(node::Node)

Return the previous node of node.

EzXML.nextelement — Function

nextelement(node::Node)

Return the next element of node.

EzXML.prevelement — Function

prevelement(node::Node)

Return the previous element of node.

EzXML.eachnode — Function

eachnode(node::Node, [backward=false])

Create an iterator of child nodes.

EzXML.nodes — Function

nodes(node::Node, [backward=false])

Create a vector of child nodes.

EzXML.eachelement — Function

eachelement(node::Node, [backward=false])

Create an iterator of child elements.

EzXML.elements — Function

elements(node::Node, [backward=false])

Create a vector of child elements.

EzXML.eachattribute — Function

eachattribute(node::Node)

Create an iterator of attributes.

eachattribute(reader::StreamReader)

Return an AttributeReader object for the current node of reader

EzXML.attributes — Function

attributes(node::Node)

Create a vector of attributes.

EzXML.hasroot — Function

hasroot(doc::Document)

Return if doc has a root element.

EzXML.hasdtd — Function

hasdtd(doc::Document)

Return if doc has a DTD node.

EzXML.hasnode — Function

hasnode(node::Node)

Return if node has a child node.

EzXML.hasnextnode — Function

hasnextnode(node::Node)

Return if node has a next node.

EzXML.hasprevnode — Function

hasprevnode(node::Node)

Return if node has a previous node.

EzXML.haselement — Function

haselement(node::Node)

Return if node has a child element.

EzXML.hasnextelement — Function

hasnextelement(node::Node)

Return if node has a next node.

EzXML.hasprevelement — Function

hasprevelement(node::Node)

Return if node has a previous node.

EzXML.hasdocument — Function

hasdocument(node::Node)

Return if node belongs to a document.

EzXML.hasparentnode — Function

hasparentnode(node::Node)

Return if node has a parent node.

EzXML.hasparentelement — Function

hasparentelement(node::Node)

Return if node has a parent node.

DOM tree modifiers

EzXML.setroot! — Function

setroot!(doc::Document, node::Node)

Set the root element of doc to node and return the root element.

EzXML.setdtd! — Function

setdtd!(doc::Document, node::Node)

Set the DTD node of doc to node and return the DTD node.

EzXML.link! — Function

link!(parent::Node, child::Node)

Link child at the end of children of parent.

EzXML.linknext! — Function

linknext!(target::Node, node::Node)

Link node as the next sibling of target.

EzXML.linkprev! — Function

linkprev!(target::Node, node::Node)

Link node as the prev sibling of target.

EzXML.unlink! — Function

unlink!(node::Node)

Unlink node from its context.

EzXML.addelement! — Function

addelement!(parent::Node, name::AbstractString)

Add a new child element of name with no content to parent and return the new child element.

addelement!(parent::Node, name::AbstractString, content::AbstractString)

Add a new child element of name with content to parent and return the new child element.

XPath query

Base.findall — Method

findall(xpath::AbstractString, doc::Document)

Find nodes matching xpath XPath query from doc.

Base.findfirst — Method

findfirst(xpath::AbstractString, doc::Document)

Find the first node matching xpath XPath query from doc.

Base.findlast — Method

findlast(doc::Document, xpath::AbstractString)

Find the last node matching xpath XPath query from doc.

Base.findall — Method

findall(xpath::AbstractString, node::Node, [ns=namespaces(node)])

Find nodes matching xpath XPath query starting from node.

The ns argument is an iterator of namespace prefix and URI pairs.

Base.findfirst — Method

findfirst(xpath::AbstractString, node::Node, [ns=namespaces(node)])

Find the first node matching xpath XPath query starting from node.

Base.findlast — Method

findlast(node::Node, xpath::AbstractString, [ns=namespaces(node)])

Find the last node matching xpath XPath query starting from node.

Validation

EzXML.validate — Function

validate(doc::Document, [dtd::Node])

Validate doc against dtd and return the validation log.

The validation log is empty if and only if doc is valid. The DTD node in doc will be used if dtd is not passed.

EzXML.readdtd — Function

readdtd(filename::AbstractString)

Read filename and create a DTD node.

Reader node types

Node type	Integer

EzXML.READER_NONE	0

EzXML.READER_ELEMENT	1

EzXML.READER_ATTRIBUTE	2

EzXML.READER_TEXT	3

EzXML.READER_CDATA	4

EzXML.READER_ENTITY_REFERENCE	5

EzXML.READER_ENTITY	6

EzXML.READER_PROCESSING_INSTRUCTION	7

EzXML.READER_COMMENT	8

EzXML.READER_DOCUMENT	9

EzXML.READER_DOCUMENT_TYPE	10

EzXML.READER_DOCUMENT_FRAGMENT	11

EzXML.READER_NOTATION	12

EzXML.READER_WHITESPACE	13

EzXML.READER_SIGNIFICANT_WHITESPACE	14

EzXML.READER_END_ELEMENT	15

EzXML.READER_END_ENTITY	16

EzXML.READER_XML_DECLARATION	17

Streaming reader

EzXML.expandtree — Method

expandtree(reader::StreamReader)

Expand the current node of reader into a full subtree that will be available until the next read of node.

Note that the expanded subtree is a read-only and temporary object. You cannot modify it or keep references to any nodes of it after reading the next node.

Currently, namespace functions and XPath query will not work on the expanded subtree.

EzXML.nodetype — Method

nodetype(reader::StreamReader)

Return the type of the current node of reader.

EzXML.nodename — Method

nodename(reader::StreamReader)

Return the name of the current node of reader.

EzXML.nodecontent — Method

nodecontent(reader::StreamReader)

Return the content of the current node of reader.

EzXML.nodedepth — Method

nodedepth(reader::StreamReader)

Return the depth of the current node of reader.

EzXML.namespace — Method

namespace(reader::StreamReader)

Return the namespace of the current node of reader.

EzXML.hasnodecontent — Method

hasnodecontent(reader::StreamReader)

Return if the current node of reader has content.

EzXML.hasnodename — Method

hasnodename(reader::StreamReader)

Return if the current node of reader has a node name.