References

Types

EzXML.Document — Type.

An XML/HTML document type.

source

EzXML.Node — Type.

A proxy type to libxml2's node struct.

source

I/O

EzXML.parsexml — Function.

parsexml(xmlstring)

Parse xmlstring and create an XML document.

source

EzXML.parsehtml — Function.

parsehtml(htmlstring)

Parse htmlstring and create an HTML document.

source

EzXML.readxml — Function.

readxml(filename)

Read filename and create an XML document.

source

EzXML.readhtml — Function.

readhtml(filename)

Read filename and create an HTML document.

source

EzXML.prettyprint — Function.

prettyprint([io], node::Node)

Print node with formatting.

source

prettyprint([io], doc::Document)

Print doc with formatting.

source

Constructors

EzXML.XMLDocument — Function.

XMLDocument(version="1.0")

Create an XML document.

source

EzXML.HTMLDocument — Function.

HTMLDocument(uri=nothing, externalID=nothing)

Create an HTML document.

source

EzXML.XMLDocumentNode — Function.

XMLDocumentNode(version)

Create an XML document node with version.

source

EzXML.HTMLDocumentNode — Function.

HTMLDocumentNode(uri, externalID)

Create an HTML document node.

uri and externalID are either a string or nothing.

source

EzXML.ElementNode — Function.

ElementNode(name)

Create an element node with name.

source

EzXML.TextNode — Function.

TextNode(content)

Create a text node with content.

source

EzXML.CommentNode — Function.

CommentNode(content)

Create a comment node with content.

source

EzXML.CDataNode — Function.

CDataNode(content)

Create a CDATA node with content.

source

EzXML.AttributeNode — Function.

AttributeNode(name, value)

Create an attribute node with name and value.

source

Node accessors

EzXML.nodetype — Method.

nodetype(node::Node)

Return the type of node as an integer.

source

EzXML.name — Method.

name(node::Node)

Return the node name of node.

source

EzXML.content — Method.

content(node::Node)

Return the node content of node.

source

EzXML.namespace — Method.

namespace(node::Node)

Return the namespace associated with node.

source

EzXML.namespaces — Method.

namespaces(node::Node)

Create a vector of namespaces applying to node.

source

EzXML.iselement — Method.

iselement(node::Node)

Return if node is an element node.

source

EzXML.isattribute — Method.

isattribute(node::Node)

Return if node is an attribute node.

source

EzXML.istext — Method.

istext(node::Node)

Return if node is a text node.

source

EzXML.iscdata — Method.

iscdata(node::Node)

Return if node is a CDATA node.

source

EzXML.iscomment — Method.

iscomment(node::Node)

Return if node is a comment node.

source

EzXML.countnodes — Method.

countnodes(parent::Node)

Count the number of child nodes of parent.

source

EzXML.countelements — Method.

countelements(parent::Node)

Count the number of child elements of parent.

source

EzXML.countattributes — Method.

countattributes(elem::Node)

Count the number of attributes of elem.

source

Node modifiers

EzXML.setname! — Method.

setname!(node::Node, name::AbstractString)

Set the name of node.

source

EzXML.setcontent! — Method.

setcontent!(node::Node, content::AbstractString)

Replace the content of node.

source

DOM tree accessors

EzXML.document — Function.

document(node::Node)

Return the document of node.

source

EzXML.root — Function.

root(doc::Document)

Return the root element of doc.

source

EzXML.parentnode — Function.

parentnode(node::Node)

Return the parent of node.

source

EzXML.firstnode — Function.

firstnode(node::Node)

Return the first child node of node.

source

EzXML.lastnode — Function.

lastnode(node::Node)

Return the last child node of node.

source

EzXML.firstelement — Function.

firstelement(node::Node)

Return the first child element of node.

source

EzXML.lastelement — Function.

lastelement(node::Node)

Return the last child element of node.

source

EzXML.nextnode — Function.

nextnode(node::Node)

Return the next node of node.

source

EzXML.prevnode — Function.

prevnode(node::Node)

Return the previous node of node.

source

EzXML.nextelement — Function.

nextelement(node::Node)

Return the next element of node.

source

EzXML.prevelement — Function.

prevelement(node::Node)

Return the previous element of node.

source

EzXML.eachnode — Function.

eachnode(node::Node)

Create an iterator of child nodes.

source

EzXML.nodes — Function.

nodes(node::Node)

Create a vector of child nodes.

source

EzXML.eachelement — Function.

eachelement(node::Node)

Create an iterator of child elements.

source

EzXML.elements — Function.

elements(node::Node)

Create a vector of child elements.

source

EzXML.eachattribute — Function.

eachattribute(node::Node)

Create an iterator of attributes.

source

EzXML.attributes — Function.

attributes(node::Node)

Create a vector of attributes.

source

DOM tree modifiers

EzXML.setroot! — Function.

setroot!(doc::Document, node::Node)

Set the root element of doc to node.

source

EzXML.link! — Function.

link!(parent::Node, child::Node)

Link child at the end of children of parent.

source

EzXML.linknext! — Function.

linknext!(target::Node, node::Node)

Link node as the next sibling of target.

source

EzXML.linkprev! — Function.

linkprev!(target::Node, node::Node)

Link node as the prev sibling of target.

source

EzXML.unlink! — Function.

unlink!(node::Ndoe)

Unlink node from its context.

source

EzXML.addelement! — Function.

addelement!(parent::Node, name::AbstractString, content::AbstractString="")

Add a new child element of name with content to parent.

source

XPath query

Base.find — Method.

find(doc::Document, xpath::AbstractString)

Find nodes matching xpath XPath query from doc.

source

Base.findfirst — Method.

findfirst(doc::Document, xpath::AbstractString)

Find the first node matching xpath XPath query from doc.

source

Base.findlast — Method.

findlast(doc::Document, xpath::AbstractString)

Find the last node matching xpath XPath query from doc.

source

Base.find — Method.

find(node::Node, xpath::AbstractString)

Find nodes matching xpath XPath query starting from node.

source

Base.findfirst — Method.

findfirst(node::Node, xpath::AbstractString)

Find the first node matching xpath XPath query starting from node.

source

Base.findlast — Method.

findlast(node::Node, xpath::AbstractString)

Find the last node matching xpath XPath query starting from node.

source

Streaming reader

EzXML.depth — Method.

depth(reader::XMLReader)

Return the depth of the current node of reader.

source

EzXML.expandtree — Method.

expandtree(reader::XMLReader)

Expand the current node of reader into a full subtree that will be available until the next read of node.

source

EzXML.nodetype — Method.

nodetype(reader::XMLReader)

Return the type of the current node of reader.

source

EzXML.name — Method.

name(reader::XMLReader)

Return the name of the current node of reader.

source

EzXML.content — Method.

content(reader::XMLReader)

Return the content of the current node of reader.

source

EzXML.namespace — Method.

namespace(reader::XMLReader)

Return the namespace of the current node of reader.

source